+++ RELEASE_NOTES.md
@@ -13,6 +13,7 @@
 ### Experimental
 
 - A new predicate, `OnlyIfPrevious`, to `filter()` messages based on the previous message.
+- A new special case of `OnlyIfPrevious`, `ChangedOnly`, to skip messages if they are equal to the previous message.
 
 ## Bug Fixes